Refact resource management:add new module mgmt to process reset and finish command.
